+++ Romawi.h
@@ -1,14 +1,15 @@
-#include "Token.h"
-#include "Bilangan.h"
 
 #ifndef _ROMAWI_H_
 #define _ROMAWI_H_
+#include "Token.h"
+#include "Bilangan.h"
+
 
 class Romawi: public Bilangan{
 	public:
 		Romawi();
 		Romawi(string);
-		int GetValue();
+		long double GetValue();
 	private:
 		int Value;
 };
